+++ packages/google_maps_flutter/google_maps_flutter/lib/google_maps_flutter.dart
@@ -28,6 +28,7 @@ export 'package:google_maps_flutter_platform_interface/google_maps_flutter_platf
         Cap,
         Circle,
         CircleId,
+        GestureHandling,
         InfoWindow,
         JointType,
         LatLng,